+++ backend/ride/src/main/java/dev/alexguesser/ride/domain/service/FareCalculator.java
@@ -2,6 +2,6 @@ package dev.alexguesser.ride.domain.service;
 
 public interface FareCalculator {
 
-    float calculate(float distance);
+    double calculate(double distance);
 
 }
