+++ backend/ride/src/main/java/dev/alexguesser/ride/domain/service/NormalFareCalculator.java
@@ -3,8 +3,8 @@ package dev.alexguesser.ride.domain.service;
 public class NormalFareCalculator implements FareCalculator {
 
     @Override
-    public float calculate(float distance) {
-        return distance * 2.1f;
+    public double calculate(double distance) {
+        return distance * 2.1;
     }
