+++ backend/ride/src/main/java/dev/alexguesser/ride/domain/service/OvernightFareCalculator.java
@@ -3,8 +3,8 @@ package dev.alexguesser.ride.domain.service;
 public class OvernightFareCalculator implements FareCalculator {
 
     @Override
-    public float calculate(float distance) {
-        return distance * 3.9f;
+    public double calculate(double distance) {
+        return distance * 3.9;
     }
 
 }
