+++ backend/ride/src/main/java/dev/alexguesser/ride/domain/service/SpecialDayFareCalculator.java
@@ -3,7 +3,7 @@ package dev.alexguesser.ride.domain.service;
 public class SpecialDayFareCalculator implements FareCalculator {
 
     @Override
-    public float calculate(float distance) {
+    public double calculate(double distance) {
         return distance;
     }
 }
