+++ backend/ride/src/main/java/dev/alexguesser/ride/infra/controller/dto/RequestRideDto.java
@@ -5,9 +5,9 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record RequestRideDto(
         String passengerId,
-        int fromLat,
-        int fromLong,
-        int toLat,
-        int toLong
+        double fromLat,
+        double fromLong,
+        double toLat,
+        double toLong
 ) {
 }
